mcp_github_integration/pyproject.toml
[project]
name = "mcp-github-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.9",
    "python-dotenv>=1.1.0",
]
